// src/app/models/year-end-inventory-models/sosTicketModel.ts
/**
 * year-end-inventory-backend
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SosTicketModel { 
    ticketNumber?: number;
    plantLocation?: number;
    storageLocation?: number;
    salesOrder?: number;
    lineItem?: number;
    material?: string;
    description?: string;
    quantity?: number;
    areaLocation?: string;
    userEntered?: string;
}